[33기 박슬기] Modify: UserData.js로 분리
