+++ src/pages/seulgi/Main/Main.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { useState } from 'react';
 import './Main.scss';
 import Nav from '../../../Components/Nav';
+import USER_LIST from './UserData';
 
 const Main = () => {
   return (
@@ -30,16 +31,6 @@ const Main = () => {
 };
 
 function FeedStory() {
-  const USER_LIST = [
-    { id: 1, userId: 'hi__yeon__' },
-    { id: 2, userId: '__plan_j' },
-    { id: 3, userId: 'suen.park' },
-    { id: 4, userId: 'jindo_luv' },
-    { id: 5, userId: 'jindo_bori' },
-    { id: 6, userId: 'poong__e' },
-    { id: 7, userId: 'peppoya' },
-    { id: 8, userId: 'jeongduck' },
-  ];
   return (
     <div className="feedStory">
       {USER_LIST.map(user => (
@@ -148,16 +139,6 @@ function Comment({ names, comments }) {
 }
 
 function RecommendBox() {
-  const USER_LIST = [
-    { id: 1, userId: 'hi__yeon__' },
-    { id: 2, userId: '__plan_j' },
-    { id: 3, userId: 'suen.park' },
-    { id: 4, userId: 'jindo_luv' },
-    { id: 5, userId: 'jindo_bori' },
-    { id: 6, userId: 'poong__e' },
-    { id: 7, userId: 'peppoya' },
-    { id: 8, userId: 'jeongduck' },
-  ];
   return (
     <div className="RightRecommendContentBox">
       {USER_LIST.map(user => (
