+++ src/pages/seulgi/Main/UserData.js
@@ -1,0 +1,12 @@
+const USER_LIST = [
+  { id: 1, userId: 'hi__yeon__' },
+  { id: 2, userId: '__plan_j' },
+  { id: 3, userId: 'suen.park' },
+  { id: 4, userId: 'jindo_luv' },
+  { id: 5, userId: 'jindo_bori' },
+  { id: 6, userId: 'poong__e' },
+  { id: 7, userId: 'peppoya' },
+  { id: 8, userId: 'jeongduck' },
+];
+
+export default USER_LIST;
